// src/selectors/funnel-sa/models/IFunnelSalesAnalys.ts
export default interface IFunnelSalesAnalys {
  readonly employeeEmail: string;
  readonly employeeID: number;
  readonly employeeName: string;
  readonly notes: string;
  readonly status: string;
  readonly stepName: string;
  readonly tanggal: string;
  readonly workflowDetailGenID: number;
  readonly workflowHeaderGenID: number;
}